requirements.txt
# Core Django
Django>=4.2,<5.0
djangorestframework>=3.14
django-environ>=0.11.0
djangorestframework-simplejwt[crypto]>=5.3.0
django-cors-headers>=4.3.0
gunicorn>=21.2.0
drf-spectacular>=0.27.0

# Database (PostgreSQL for Render)
psycopg2-binary>=2.9.9
dj-database-url>=2.1.0

# Static files
whitenoise>=6.6.0

# Celery (required - used in code, but runs in eager mode)
celery>=5.3.0
django-celery-beat>=2.5.0
django-redis>=5.4.0

# Storage (optional)
django-storages>=1.14.0
boto3>=1.28.0

# Admin theme
django-unfold>=0.30.0

# SMS (optional)
twilio>=8.0.0

# Error tracking (optional)
sentry-sdk>=1.39.0
